Update setDefaultGroupandDistro.ps1
Redacted any company info and upgraded script to include more error checking and the continuOn function to ask if a user wanted to start the script again. This helps if there is a one off that the script runner will need to add after the script is finished. By just looping through the script (passed the Connect-ExchangeOnline) they won't have to take the time to re-enter their credentials.